MS6 Salvation
Currently the call to ENES100 in the setup crashes the program. I don't know why ENES100.begin fails, you have to change the value of the room number for testing. If you want to activate a specific function, just call that one in the loop and comment everything else out.